Use fcntl, instead of the contents of the pid file, to get the PID of the process holding the lock on the pid file.
It's not that I don't like pid files, but it does seem like an ugly way
to get a best guess at something (the pid of a process with a lock) when
there are standard calls to get the same information precisely. In
particular, while there's no guarantee that the process with the pid in
the pidfile actually has the lock, or even is the process we want it to
be, if fcntl returns a pid then we know with certainty that that process
does (or did) have the lock.