Drop the HTTP server/client; use json over a socket instead.
This is supposed to be simple and low-overhead, once we got to the point
where writing it made sense to write a custom client to be able to
control our cl-service service, well that's what we'd gone too far.

This change drops the monkey patching on hunchentoot and usocket, drop
the local-socket-acceptor entirely, drops all the files except
cl-service.lisp (and the package.lisp file).

Now we just have a small server which reads one json array and returns a
json object:

  $ echo "[\"status\"]" | socat unix-socket:./control.socket -

no http header parsing (there just aren't many http clients that will
talk to a domain socket), it's just simple json in and json out.